Fix compilation error

+++ GasolinerasVIP/Views/GasStationPage.xaml.cs
@@ -14,9 +14,9 @@ namespace GasolinerasVIP.Views
     [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class GasStationPage : ContentPage
     {
-        private double TotalMagna   = 0.0;
-        private double TotalPremium = 0.0;
-        private double Total = 0.0;
+        private decimal TotalMagna   = 0.0m;
+        private decimal TotalPremium = 0.0m;
+        private decimal Total = 0.0m;
         GasStation gasStation;
         public GasStationPage(GasStation gasStation)
         {
@@ -31,29 +31,29 @@ namespace GasolinerasVIP.Views
             MagnaPrice.Text   = String.Format("${0:f2}", gasStation.MagnaPrice);
             PremiumPrice.Text = String.Format("${0:f2}" ,gasStation.PremiumPrice);
             
-            TotalMagna   = Double.Parse(MagnaQt.Text) * gasStation.MagnaPrice;
-            TotalPremium = Double.Parse(PremiumQt.Text) * gasStation.PremiumPrice;
+            TotalMagna   = Decimal.Parse(MagnaQt.Text) * gasStation.MagnaPrice;
+            TotalPremium = Decimal.Parse(PremiumQt.Text) * gasStation.PremiumPrice;
 
             TotalMagnaPrice.Text   = String.Format("${0:f2}", TotalMagna);
             TotalPremiumPrice.Text = String.Format("${0:f2}", TotalPremium);
 
-            Total = TotalMagna + TotalPremium + 30.00;
+            Total = TotalMagna + TotalPremium + 30.00m;
             TotalLbl.Text = String.Format("${0:f2}", Total);
         }
         void Magna_Entry(object sender, EventArgs e)
         {
             MagnaQtDtl.Text = MagnaQt.Text;
-            TotalMagna = Double.Parse(MagnaQt.Text) * gasStation.MagnaPrice;
+            TotalMagna = Decimal.Parse(MagnaQt.Text) * gasStation.MagnaPrice;
             TotalMagnaPrice.Text = String.Format("${0:f2}", TotalMagna);
-            Total = TotalMagna + TotalPremium + 30.00;
+            Total = TotalMagna + TotalPremium + 30.00m;
             TotalLbl.Text = String.Format("${0:f2}", Total);
         }
         void Premium_Entry(object sender, EventArgs e)
         {
             PremiumQtDtl.Text = PremiumQt.Text;
-            TotalPremium = Double.Parse(PremiumQt.Text) * gasStation.PremiumPrice;
+            TotalPremium = Decimal.Parse(PremiumQt.Text) * gasStation.PremiumPrice;
             TotalPremiumPrice.Text = String.Format("${0:f2}", TotalPremium);
-            Total = TotalMagna + TotalPremium + 30.00;
+            Total = TotalMagna + TotalPremium + 30.00m;
             TotalLbl.Text = String.Format("${0:f2}", Total);
         }
         private async void Btn_RealizarCompra(object sender, EventArgs e)
@@ -64,8 +64,8 @@ namespace GasolinerasVIP.Views
                 Id= 1,
                 userId= 1,
                 Station = gasStation,
-                OrderedMagna = Double.Parse(MagnaQt.Text),
-                OrderedPremium = Double.Parse(PremiumQt.Text),
+                OrderedMagna = Decimal.Parse(MagnaQt.Text),
+                OrderedPremium = Decimal.Parse(PremiumQt.Text),
                 state= 1,
             };
             await Navigation.PushAsync(new PaymentPage(myOrder));
